M2/app/requirements.txt
numpy
pandas
PyYAML
requests
SQLAlchemy
psycopg2-binary
bs4
scikit-learn
scipy
docker
kafka-python